pyproject.toml
[project]
name = "earnest-rcm-ai-coding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-lib>=2.225.0",
    "boto3==1.34.0",
    "django>=5.0.0",
    "djangorestframework>=3.15.0",
    "fastapi==0.112.2",
    "ipykernel>=7.1.0",
    "ipython>=9.7.0",
    "ipython-sql>=0.5.0",
    "langchain>=1.1.0",
    "langchain-openai>=1.1.0",
    "langchain-postgres>=0.0.16",
    "langchain-chroma>=0.2.3",
    "matplotlib>=3.10.7",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    "openpyxl>=3.0.0",
    "pandas>=1.5.0",
    "pgvector>=0.3.6",
    "plotly>=6.5.0",
    "psycopg2-binary>=2.9.11",
    "pydantic==2.8.2",
    "pymupdf>=1.24.5",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.9",
    "sentence-transformers>=5.1.0",
    "tiktoken>=0.12.0",
    "uvicorn[standard]==0.30.6",
    "scikit-learn>=1.7.2",
]
